backend/requirements.txt
fastapi[standard]>=0.115.12,<0.116.0
uvicorn[standard]>=0.32.0,<0.33.0
python-multipart>=0.0.12
python-docx>=1.1.2
reportlab>=4.2.5
language-tool-python==2.9.4
textract==1.6.3
pydantic>=2.9.2,<3.0.0
python-jose[cryptography]>=3.3.0
aiofiles>=24.1.0
spacy>=3.7.0,<4.0.0
nltk>=3.9.1
